+++ src/test/fulcro/tests_to_run.cljs
@@ -10,6 +10,7 @@
     fulcro.client.routing-spec
     fulcro.client.util-spec
     fulcro.ui.forms-spec
+    fulcro.server-render-spec
     fulcro.i18n-spec))
 
 ;********************************************************************************